REFAC: first steps into refactoring client to ts non-strict
