+++ client/package.json
@@ -52,6 +52,7 @@
     ]
   },
   "devDependencies": {
+    "@types/reach__router": "^1.3.7",
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.2",
     "typescript": "^4.2.3"
