+++ client/src/components/Home/Home.tsx
@@ -3,8 +3,15 @@ import Header from './Header/Header';
 import ServicesList from './ServicesList/ServicesList';
 import BookingSteps from './BookingSteps/BookingSteps';
 import Testimonial from './Testimonial/Testimonial';
+import {  RouteComponentProps } from "@reach/router"
 
-const Home = ({ user }) => {
+interface HomeProps {
+	user: object;
+	path: RouteComponentProps;
+}
+
+
+const Home:React.FC<HomeProps> = ({ user }) => {
 	return (
 		<>
 			<Header user={user} />
