+++ client/src/components/Profile/Profile.tsx
@@ -1,6 +1,13 @@
+import {  RouteComponentProps } from "@reach/router"
 import './Profile.css';
 
-const Profile = () => {
+
+interface ProfileProps {
+	path: RouteComponentProps;
+}
+
+
+const Profile: React.FC<ProfileProps> = () => {
 	return (
 		<div className="profile-page">
 			<h1>Profile Page </h1>
